+++ src/openrct2/entity/JumpingFountain.h
