+++ src/openrct2/libopenrct2.vcxproj
@@ -281,7 +281,7 @@
     <ClInclude Include="entity\EntityList.h" />
     <ClInclude Include="entity\EntityRegistry.h" />
     <ClInclude Include="entity\EntityTweener.h" />
-    <ClInclude Include="entity\Fountain.h" />
+    <ClInclude Include="entity\JumpingFountain.h" />
     <ClInclude Include="entity\Guest.h" />
     <ClInclude Include="entity\Litter.h" />
     <ClInclude Include="entity\MoneyEffect.h" />
@@ -400,7 +400,7 @@
     <ClInclude Include="paint\Painter.h" />
     <ClInclude Include="paint\Balloon.h" />
     <ClInclude Include="paint\entity\Paint.Duck.h" />
-    <ClInclude Include="paint\entity\Paint.Fountain.h" />
+    <ClInclude Include="paint\entity\Paint.JumpingFountain.h" />
     <ClInclude Include="paint\entity\Paint.Guest.h" />
     <ClInclude Include="paint\entity\Paint.Litter.h" />
     <ClInclude Include="paint\entity\Paint.Peep.h" />
@@ -905,7 +905,7 @@
     <ClCompile Include="entity\EntityBase.cpp" />
     <ClCompile Include="entity\EntityRegistry.cpp" />
     <ClCompile Include="entity\EntityTweener.cpp" />
-    <ClCompile Include="entity\Fountain.cpp" />
+    <ClCompile Include="entity\JumpingFountain.cpp" />
     <ClCompile Include="entity\Guest.cpp" />
     <ClCompile Include="entity\Litter.cpp" />
     <ClCompile Include="entity\MoneyEffect.cpp" />
@@ -1001,7 +1001,7 @@
     <ClCompile Include="paint\PaintHelpers.cpp" />
     <ClCompile Include="paint\entity\Paint.Balloon.cpp" />
     <ClCompile Include="paint\entity\Paint.Duck.cpp" />
-    <ClCompile Include="paint\entity\Paint.Fountain.cpp" />
+    <ClCompile Include="paint\entity\Paint.JumpingFountain.cpp" />
     <ClCompile Include="paint\entity\Paint.Guest.cpp" />
     <ClCompile Include="paint\entity\Paint.Litter.cpp" />
     <ClCompile Include="paint\entity\Paint.Peep.cpp" />
