+++ src/openrct2/paint/Paint.Entity.cpp
@@ -13,7 +13,7 @@
 #include "../drawing/LightFX.h"
 #include "../entity/Duck.h"
 #include "../entity/EntityList.h"
-#include "../entity/Fountain.h"
+#include "../entity/JumpingFountain.h"
 #include "../entity/MoneyEffect.h"
 #include "../entity/Particle.h"
 #include "../entity/Staff.h"
@@ -28,8 +28,8 @@
 #include "Paint.h"
 #include "entity/Paint.Balloon.h"
 #include "entity/Paint.Duck.h"
-#include "entity/Paint.Fountain.h"
 #include "entity/Paint.Guest.h"
+#include "entity/Paint.JumpingFountain.h"
 #include "entity/Paint.Litter.h"
 #include "entity/Paint.Staff.h"
 #include "vehicle/VehiclePaint.h"
