+++ src/openrct2/paint/entity/Paint.JumpingFountain.h
