+++ src/openrct2/park/ParkFile.cpp
@@ -31,7 +31,7 @@
 #include "../entity/Duck.h"
 #include "../entity/EntityList.h"
 #include "../entity/EntityRegistry.h"
-#include "../entity/Fountain.h"
+#include "../entity/JumpingFountain.h"
 #include "../entity/Litter.h"
 #include "../entity/MoneyEffect.h"
 #include "../entity/Particle.h"
