+++ src/openrct2/rct2/S6Importer.cpp
@@ -17,8 +17,8 @@
 #include "../core/String.hpp"
 #include "../entity/Balloon.h"
 #include "../entity/Duck.h"
-#include "../entity/Fountain.h"
 #include "../entity/Guest.h"
+#include "../entity/JumpingFountain.h"
 #include "../entity/Litter.h"
 #include "../entity/MoneyEffect.h"
 #include "../entity/Particle.h"
