+++ src/openrct2/scripting/bindings/world/ScMap.cpp
@@ -15,8 +15,8 @@
     #include "../../../entity/Balloon.h"
     #include "../../../entity/Duck.h"
     #include "../../../entity/EntityList.h"
-    #include "../../../entity/Fountain.h"
     #include "../../../entity/Guest.h"
+    #include "../../../entity/JumpingFountain.h"
     #include "../../../entity/Litter.h"
     #include "../../../entity/MoneyEffect.h"
     #include "../../../entity/Particle.h"
