+++ src/openrct2/world/Map.cpp
@@ -27,7 +27,7 @@
 #include "../entity/Duck.h"
 #include "../entity/EntityList.h"
 #include "../entity/EntityTweener.h"
-#include "../entity/Fountain.h"
+#include "../entity/JumpingFountain.h"
 #include "../entity/PatrolArea.h"
 #include "../entity/Staff.h"
 #include "../interface/Cursors.h"
